* Added logging of size of written data to hard drive.

+++ src/rtp_foutput.h
@@ -91,6 +91,6 @@ int rtp_close_stream_output(struct rtp_stream *stream);
  *\param len Length of the packet.
  *\param stream Stream that packet belongs to.
  */
-int rtp_write_packet(rtp_session_type_t stream_type, RD_buffer_t *packet, int len, struct rtp_stream *stream);
+ssize_t rtp_write_packet(rtp_session_type_t stream_type, RD_buffer_t *packet, int len, struct rtp_stream *stream);
 
 #endif /* RTP_FOUTPUT_H_ */
